cleaning up + envs in separated file
