+++ src/test/5.customActions.test.ts
@@ -1,8 +1,9 @@
 import { By, CustomTreeSection, EditorView, InputBox, until, VSBrowser, WebDriver, Workbench} from 'vscode-extension-tester';
 import { expect } from 'chai';
 import { delay, initialize, validateRootNode, validateSeverities } from './utils/utils';
-import { CX_CLEAR, CX_FILTER_CONFIRMED, CX_FILTER_HIGH, CX_FILTER_INFO, CX_FILTER_LOW, CX_FILTER_MEDIUM, CX_FILTER_NOT_EXPLOITABLE, CX_FILTER_NOT_IGNORED, CX_FILTER_PROPOSED_NOT_EXPLOITABLE, CX_FILTER_TO_VERIFY, CX_FILTER_URGENT, CX_GROUP_FILE, CX_GROUP_LANGUAGE, CX_GROUP_QUERY_NAME, CX_GROUP_STATE, CX_GROUP_STATUS, CX_LOOK_SCAN, SAST_TYPE, SCAN_KEY_TREE } from './constants';
+import { CX_CLEAR, CX_FILTER_CONFIRMED, CX_FILTER_HIGH, CX_FILTER_INFO, CX_FILTER_LOW, CX_FILTER_MEDIUM, CX_FILTER_NOT_EXPLOITABLE, CX_FILTER_NOT_IGNORED, CX_FILTER_PROPOSED_NOT_EXPLOITABLE, CX_FILTER_TO_VERIFY, CX_FILTER_URGENT, CX_GROUP_FILE, CX_GROUP_LANGUAGE, CX_GROUP_QUERY_NAME, CX_GROUP_STATE, CX_GROUP_STATUS, CX_LOOK_SCAN, SAST_TYPE, SCAN_KEY_TREE } from './utils/constants';
 import { waitByLinkText } from './utils/waiters';
+import { SCAN_ID } from './utils/envs';
 
 describe('filter and groups actions tests', () => {
 	let bench:Workbench;
@@ -29,20 +30,20 @@ describe('filter and groups actions tests', () => {
 		await bench.executeCommand(CX_LOOK_SCAN);
 		let input = await InputBox.create();
 		await input.setText(
-			process.env.CX_TEST_SCAN_ID
+			SCAN_ID
 		);
 		await input.confirm();
-		await waitByLinkText(driver, SCAN_KEY_TREE + process.env.CX_TEST_SCAN_ID, 15000);
+		await waitByLinkText(driver, SCAN_KEY_TREE + SCAN_ID, 15000);
 		const commands = [{command:CX_FILTER_INFO,text:"INFO"},{command:CX_FILTER_LOW,text:"LOW"},{command:CX_FILTER_MEDIUM,text:"MEDIUM"},{command:CX_FILTER_HIGH,text:"HIGH"}];
 		for (var index in commands) {
 			await bench.executeCommand(commands[index].command);
 			treeScans = await initialize();
 			let scan = await treeScans?.findItem(
-				SCAN_KEY_TREE + process.env.CX_TEST_SCAN_ID
+				SCAN_KEY_TREE + SCAN_ID
 			);
 			while(scan===undefined){
 				scan = await treeScans?.findItem(
-					SCAN_KEY_TREE +process.env.CX_TEST_SCAN_ID
+					SCAN_KEY_TREE +SCAN_ID
 				);
 			}
 			let isValidated = await validateSeverities(scan, commands[index].text);
@@ -59,11 +60,11 @@ describe('filter and groups actions tests', () => {
 		// Get scan node
 		const treeScans = await initialize();
 		let scan =  await treeScans?.findItem(
-			SCAN_KEY_TREE + process.env.CX_TEST_SCAN_ID
+			SCAN_KEY_TREE + SCAN_ID
 		);
 		while(scan===undefined){
 			scan = await treeScans?.findItem(
-				SCAN_KEY_TREE + process.env.CX_TEST_SCAN_ID
+				SCAN_KEY_TREE + SCAN_ID
 				);
 		}
 		// Expand and validate scan node to obtain engine nodes
