+++ src/test/6.scan.test.ts
@@ -9,8 +9,9 @@ import {
 } from "vscode-extension-tester";
 import { expect } from "chai";
 import { delay, initialize } from "./utils/utils";
-import { CX_CLEAR, CX_LOOK_SCAN, SCAN_KEY_TREE, VS_OPEN_FOLDER } from "./constants";
+import { CX_CLEAR, CX_LOOK_SCAN, SCAN_KEY_TREE, VS_OPEN_FOLDER } from "./utils/constants";
 import { waitByLinkText } from "./utils/waiters";
+import { SCAN_ID } from "./utils/envs";
 
 describe("Scan from IDE", () => {
   let bench: Workbench;
@@ -34,16 +35,16 @@ describe("Scan from IDE", () => {
     await  bench.executeCommand(CX_LOOK_SCAN);
     let input = await InputBox.create();
     await input.setText(
-      process.env.CX_TEST_SCAN_ID 
+      SCAN_ID 
     );
     await input.confirm();
-    await waitByLinkText(driver,  SCAN_KEY_TREE + process.env.CX_TEST_SCAN_ID, 5000);
+    await waitByLinkText(driver,  SCAN_KEY_TREE + SCAN_ID, 5000);
     let scan = await treeScan?.findItem(
-      SCAN_KEY_TREE + process.env.CX_TEST_SCAN_ID 
+      SCAN_KEY_TREE + SCAN_ID 
     );
     while (scan === undefined) {
       scan = await treeScan?.findItem(
-        SCAN_KEY_TREE + process.env.CX_TEST_SCAN_ID 
+        SCAN_KEY_TREE + SCAN_ID 
       );
     }
     // click play button(or initiate scan with command)
