+++ src/test/utils/envs.ts
@@ -1,0 +1,3 @@
+export const API_KEY:string = process.env.CX_API_KEY;
+export const SCAN_ID:string =process.env.CX_TEST_SCAN_ID;
+export const CX_TEST_SCAN_PROJECT_NAME: string = process.env.CX_TEST_PROJECT_NAME ? process.env.CX_TEST_PROJECT_NAME : "webgoat";
